Change Long To BigInteger

+++ src/main/java/pl/coderslab/library/Service/BookRepository.java
@@ -1,9 +1,10 @@
 package pl.coderslab.library.Service;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import pl.coderslab.library.Model.Book;
 
-import java.awt.print.Book;
+import java.math.BigInteger;
 
-public interface BookRepository extends JpaRepository<Book, Long> {
-    public Book findBookByIsbn(Long isbn);
+public interface BookRepository extends JpaRepository<Book, BigInteger> {
+    public Book findBookByIsbn(BigInteger isbn);
 }
